fixes
fixed frustum precision.
dynamic smoothing for light distances in light compute shader

some experimental changes with fractional offset (world scrolling)

fixes for gui drawing

+++ MinCity/cNuklear.cpp
@@ -1684,7 +1684,7 @@ void  cNuklear::UpdateGUI()
 			if (bNewCity) {
 				
 				std::string const szCityName( szEdit );
-				size_t const checkedLength( SFM::min(szCityName.length(), (size_t)szLength) );
+				size_t const checkedLength( SFM::min((int32_t)szCityName.length(), szLength) );
 
 				nk_text_animated(_ctx, s, 0.3f, _fonts[eNK_FONTS::SMALL], _fonts[eNK_FONTS::ENCRYPT_SMALL], _saveWindow.thumbnail_text_state,
 					szCityName.substr(0, checkedLength),
